requirements.txt
google-generativeai>=0.8.0
google-genai>=1.0.0
streamlit>=1.35.0
python-dotenv>=1.0.0
